+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ### Added
 
 - Add item_setter expression
+- Add PythonStatement
+
+### Changed
+
+- Make it clear if we are import a module or an attribute of a module